Code refactoring and namespaces
