+++ Assets/Scripts/Tank.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6d39fea8026ca704fa4536bd8e54e6f4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
